compute tclhandleEntryAlignment at compile time
There is no need to defer this to runtime. Making this a compile-time constant
allows the compiler to constant-fold computations involving it, producing more
efficient code.

Note, this is not really even the correct notion of platform alignment. The
alignment of a type is not guaranteed to be the same as its size. Nevertheless,
we leave this alone for now.