remove test suite assumption that Python can be run as `python3`
On Windows, python3.exe does not typically exist. The python.exe binary is
Python 3 and is intended to be called as just this. The CI setup currently works
around this by copying python.exe to python3.exe. This change removes any
dependency on python3 existing, which should ease life for both CI and Windows
users wanting to run the test suite. Closes #2049.