remove python3.exe copying in Windows CI
With the prior couple of changes, neither the Windows build nor the test suite
relies on a python3.exe binary. So we can simplify Windows CI by removing these
steps, leaving only a python.exe binary. This is more reflective of the typical
environment a user will have too. Related to #2049.