remove use of double literals with ydelta in make_flat_labeled_edge
Commit ebe7517ae4eb86ee4bf34e1af31a174f1f0a4049 altered these lines to use
double literals instead of integer literals for `5` and `6`. It is not clear why
it did this as ydelta is an integer variable. The effect of that change was to
ask the compiler to perform the computation itself using doubles and then reduce
it to an integer result. This commit reverts that part of the diff to simply use
integer computation throughout. This change was motivated by the compiler
diagnosing this issue with -Wfloat-conversion.