fix: remove dangling reference to textfont.* in Autotools file
The file textfont.c was renamed to textspan.c in
65f337664d177b588c5c2418fe12bac7eef62c2a but this rule was incorrectly not
updated. Rather than relying on manual dependencies which the Autotools docs
suggest can be problematic,¹ we just remove this rule. This removes the ability
to `make textspan.o`, but clearly no one was relying on this anyway as it has
been broken since 2013.

¹ https://www.gnu.org/software/automake/manual/html_node/Built-Sources-Example.html