remove work around for stateful macOS CI runners
Graphviz was part of a closed beta program for Gitlab macOS CI runners.¹ During
the beta period, Graphviz had access to a single macOS VM environment that was
stateful. That is, the effects of any macOS CI task would persist to the next
macOS CI task. To work around this, logic was introduced to manually remove the
Graphviz installed in the macOS CI environment by the prior run.

The Gitlab program will transition into open beta on 2021-05-22, however the
Graphviz configuration was already migrated to the open beta model on
2021-04-07.² This means the macOS runner are no longer stateful:

  To recap, for the closed beta, we provisioned a static macOS virtual machine
  for each project. This meant that we executed any pipeline job in your
  associated repository on the same virtual machine each time. Since the VM was
  dedicated to your project, this also meant that your pipeline job did not have
  to wait in a queue before running.

  For the open-beta, each job that requires a macOS build environment will run
  in its own ephemeral virtual machine provisioned on demand by the GitLab
  Runner macOS autoscaler. This means that you can’t rely on persistent storage
  anymore.

This change removes the work around that is no longer needed. This reverts
62165551efb389ea14a6709ffd89b2b5ed79e35c.

  ¹ https://gitlab.com/gitlab-com/macos-buildcloud-runners-beta/-/issues/3
  ² https://gitlab.com/gitlab-com/macos-buildcloud-runners-beta/-/issues/3#note_536852935