optimize 0 case in xdot_fmt_num
Commit 894e94c66340187ff6f8ea97c9336ea048a3d6ad applied a performance
optimization to gvprintdouble that was contributing to a performance issue in a
user’s graph. The code touched in this commit has the same pattern as that in
gvprintdouble, and the same corresponding (latent) performance issue. While it
is not known to cause any problems currently, we may as well proactively apply
the same optimization here.

This preserves the same functionality, but simply takes an early exit from the
function when we already know precisely how to print the resulting string.