sfio: remove uwin specific handling
The indirection of the data through the IAT is implicitly handled by
modern linkers and compilers when the data symbol is annotated with
`__declspec(dll{import,export})`.  This removes the support for the
older toolchains as part of the work to properly annotate symbols for
improvements to the Windows port.