add a unified XML escaping function
This change adds to cgraph a function that performs a superset of current XML
escaping done within Graphviz. This is intended to be an all-singing,
all-dancing alternative that subsumes and deprecates all other existing XML
escaping functions. As horrifying as that sentence sounds, this is believed to
be more reasonable than the current duplication of XML escaping functionality.

In particular, this function subsumes the following:

  * cmd/tools/gv2gxl.c:xml_string
  * cmd/smyrna/smyrna_utils.c:xml_string
  * lib/common/labels.c:xml_string
  * plugin/core/gvrender_core_vml.c:html_string

A future change will remove the above and replace them all with calls to
gv_xml_esc.

This new function differs from the above in a few notable ways:

  1. It attempts more strict conformance with the XML specification. In
     particular, it has a more precise definition of what can be an XML entity
     name. The hope is to avoid latent issues like, e.g., html_string above does
     not recognize “sup2” as an entity name.

  2. It handles UTF-8 more rigorously. The html_string implementation both
     arbitrarily aborts in some cases that appear legal and also is too liberal
     in accepting malformed UTF-8.

  3. It does not recognize “&#;” or “&#x;” as XML entities. The previous escape
     functions do, but this appears to be accidental rather than deliberate.

Though it is exported from cgraph, its header is not shipped. It is assumed
that users have no need to XML-escape strings or have better ways of doing this,
so the header would just clutter their include directories. Nevertheless the
function gv_xml_esc needs to be exported to be usable by other Graphviz
components.

Towards #1868.