compute the extent of some gvputs writes at compile time
This change teaches the compiler how to unfold gvputs of a string literal into
gvwrite, computing the length of the literal at compile time. This is only
applied to the SVG back end in this commit.

On tests/regression_tests/large/long_chain, this drops the number of gvputs
calls from 2442066 to 363011, though obviously introduces the difference as
gvwrite calls. The total executed instructions drops from 8093310656 to
8009099650, a speed up of ~1%.