more portable exit status from nop
The exit status of nop was a bitwise-OR of two error counts. I do not see how
the actual value of this could ever be useful, except for being non-zero.
However, a non-zero exit status is not always an error. E.g. on VMS there are
some non-zero exit statuses that still indicate success.

This change causes nop to more reliably and portably exit with EXIT_FAILURE when
encountering errors.