fix: do not fclose(NULL) in gvmap
gvmap opens a debugging file to write timing information to if verbosity is
enabled. It tried to unconditionally close this file, regardless of whether
verbosity was enabled. The result was that, without verbosity enabled, it would
call fclose(NULL). The effect of this seems to vary across operating systems but
can result in a crash. This fix only tries to close the file if it was
previously opened. Related to #1822.

Unfortunately we cannot un-xfail test_gvmap_fclose because it fails for another
reason on some platforms.