add a macro to be used when declaring external functions in ortho
The macro expands to import/export declarations on Windows that are
needed for dynamic-link libraries (DLLs) and nothing for other OSes.

This is a preparation for being able to include the ortho object files
into the gvc shared/dynamic-link library in an upcoming commit in this
series.

Note that this change continues to use the anti-pattern of GVC_EXPORTS
which is not automatically generated by CMake like gvc_EXPORTS is.

Towards https://gitlab.com/graphviz/graphviz/-/issues/2096.