return a value instead of pointer from Blocks::totalOrder
This function dynamically allocated a list and returned a pointer to this,
presumably under the assumption that copying this kind of object during return
is expensive. However, Named Return Value Optimization (NRVO) applies to this
function. So the compiler can elide the copy during return, making a value
cheaper than a heap pointer here.