remove pow micro-optimizations in spring_electrical.c
This code was attempting to optimize the case when the second argument to pow()
was 2. However, modern compilers are able to understand such nuances. In fact,
at -O3 recent Clang *undoes* this branching code into an unconditional call to
pow(). In other words, it considers this branch to multiplication to be a
de-optimization.

This change also incidentally removes 14 -Wfloat-equal warnings.