remove legacy __eprintf dependency check
This check was forcing assertions off if using them depends on __eprintf. It is
unclear to me what the motivation for this is and it was introduced in the
initial revision with no further information. As far as I am aware, this is no
longer relevant. Modern libc is intended to provide everything necessary to
support assert, unless you are building freestanding which Graphviz does not
support.