remove Courier.gv graph reference test
This test compares the rendering of a graph using the Courier font against
reference examples. However the rendering of Courier is not stable across
operating system releases. At present, this test is spuriously failing due to a
macOS upgrade. But in principle, an upgrade of any platform can break it. To
avoid testing this unstable behavior, we remove this test.