remove useless MSBuild project reference to ortho for dot
dot only depends indirectly on ortho through dotgen which it already
has a project reference to.

This change is made to avoid confusion when upcoming commits in this
series change ortho to be part of the gvc shared library and adjusts
the dotgen library accordingly.

Towards https://gitlab.com/graphviz/graphviz/-/issues/2096.