remove useless MSBuild project dependency on ortho for neato_layout plugin
The neato_layout plugin only depends indirectly on ortho through
neatogen which it already has a project reference to.

A project reference is a more modern way of creating a project
dependency than the centralized project dependencies in the solution
file. See https://stackoverflow.com/a/2317017/3122101.

This change is made to avoid confusion when upcoming commits in this
series change ortho to be part of the gvc shared library and adjusts
the neatogen library accordingly.

Towards https://gitlab.com/graphviz/graphviz/-/issues/2096.