remove useless MSBuild project dependency on ortho for gvmap
gvmap only depends indirectly on ortho through neatogen which it
already has a project dependency on.

This change is made to avoid confusion when upcoming commits in this
series change ortho to be part of the gvc shared library and adjusts
the neatogen library accordingly.

Towards https://gitlab.com/graphviz/graphviz/-/issues/2096.