replace mkdefs-generated gdefs.h with X macros
Prior to this commit, the gdefs.h header was generated by a C program, mkdefs.c.
There were a number of issues with this approach:

  1. The CMake build system was assuming the compiler to build mkdefs.c and the
     compiler to build Graphviz itself were the same. This is not necessarily
     true when cross-compiling.

  2. Generation under MSBuild seems to have been impractical, so the generated
     header was checked in to the repository under windows/include/gvpr/gdefs.h,
     somewhat defeating the purpose of making it generated.

  3. The CMake build system seems to not have been setup to correctly compile
     mkdefs.c under all circumstances (see #2101).

This change removes any reliance on a host C compiler and instead uses a series
of X macros¹ to achieve the same effect. The values of all generated constants
and the content of generated structures is intended to be unchanged, though some
#defines have been altered to enums. In these cases, there was no advantage to
using a macro and multiple advantages to not using a macro.

This change is affecting a shipped header (gdefs.h) and also removes it from the
list of shipped headers. Installing it appears to have been a mistake as there
is no easy way for end users to use it. The header, fully expanded, still relies
on further expansion of macros that are only defined in expr.h, a header that is
not shipped.

Closes #2101.

¹ https://en.wikipedia.org/wiki/X_Macro