implement manual comparison and hashing functions for reference counted strings
In this change, comparison and hashing functions that were previously defaulted
to `strcmp` and `dtstrhash` are now implemented explicitly for the reference
counted strings dictionary. These explicit implementations do exactly the same
thing as the previous implicit defaults, using `strcmp` and `dtstrhash`.

OK, so if they are identical to the defaults, why implement them at all? There
exists a current bug where regular strings and HTML-like strings with the same
content cannot coexist in a reference counted string dictionary (#2089).
Introducing explicit key handling functions is the first step to resolving this.